requirements.txt
# VELOCITY-ASR v2 Dependencies
# ============================

# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0

# ONNX export and runtime
onnx>=1.14.0
onnxruntime>=1.15.0

# Configuration
pyyaml>=6.0

# Progress bars (for scripts)
tqdm>=4.65.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.15.0

# Optional: Additional audio formats support
# soundfile>=0.12.0
# librosa>=0.10.0

# Development dependencies (uncomment for development)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# isort>=5.12.0
# flake8>=6.0.0
# mypy>=1.0.0
